feat!: swipe to switch between tabs
