+++ app/src/main/kotlin/com/aistra/hail/ui/home/HomeAdapter.kt
@@ -1,94 +1,11 @@
 package com.aistra.hail.ui.home
 
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.TextView
-import androidx.recyclerview.widget.DiffUtil
-import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
-import com.aistra.hail.R
-import com.aistra.hail.app.AppInfo
-import com.aistra.hail.app.HailData
-import com.aistra.hail.utils.AppIconCache
-import com.aistra.hail.utils.HPackages.myUserId
-import com.google.android.material.color.MaterialColors
-import kotlinx.coroutines.Job
+import androidx.fragment.app.Fragment
+import androidx.viewpager2.adapter.FragmentStateAdapter
 
-object HomeAdapter : ListAdapter<AppInfo, HomeAdapter.ViewHolder>(HomeDiff()) {
-    var multiselect: Boolean = false
-    val selectedList = mutableListOf<AppInfo>()
-    private var loadIconJob: Job? = null
-    lateinit var onItemClickListener: OnItemClickListener
-    lateinit var onItemLongClickListener: OnItemLongClickListener
+class HomeAdapter(fragment: HomeFragment, val fragments: MutableList<PagerFragment>) :
+    FragmentStateAdapter(fragment) {
+    override fun getItemCount(): Int = fragments.size
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder = ViewHolder(
-        LayoutInflater.from(parent.context).inflate(R.layout.item_home, parent, false)
-    )
-
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        val info = currentList[position]
-        holder.itemView.run {
-            setOnClickListener { onItemClickListener.onItemClick(info) }
-            setOnLongClickListener { onItemLongClickListener.onItemLongClick(info) }
-            info.selected = info.isNowSelected(selectedList)
-            info.state = info.getCurrentState()
-            findViewById<ImageView>(R.id.app_icon).run {
-                info.applicationInfo?.let {
-                    loadIconJob = AppIconCache.loadIconBitmapAsync(
-                        context,
-                        it,
-                        myUserId,
-                        this,
-                        HailData.grayscaleIcon && info.state == AppInfo.STATE_FROZEN
-                    )
-                } ?: run {
-                    setImageDrawable(context.packageManager.defaultActivityIcon)
-                    colorFilter = null
-                }
-            }
-            findViewById<TextView>(R.id.app_name).run {
-                text = buildString {
-                    if (!HailData.grayscaleIcon && info.state == AppInfo.STATE_FROZEN) append("❄️")
-                    if (info.whitelisted) append("\uD83D\uDD12")
-                    append(info.name)
-                }
-                isEnabled = !HailData.grayscaleIcon || info.state != AppInfo.STATE_FROZEN
-                when {
-                    info.selected -> setTextColor(
-                        MaterialColors.getColor(this, R.attr.colorPrimary)
-                    )
-                    info.state == AppInfo.STATE_UNKNOWN -> setTextColor(context.getColorStateList(R.color.color_warn))
-                    else -> setTextAppearance(R.style.TextAppearance_Material3_BodyMedium)
-                }
-            }
-        }
-    }
-
-    fun onDestroy() {
-        multiselect = false
-        selectedList.clear()
-        if (loadIconJob?.isActive == true) loadIconJob?.cancel()
-    }
-
-    class HomeDiff : DiffUtil.ItemCallback<AppInfo>() {
-        override fun areItemsTheSame(oldItem: AppInfo, newItem: AppInfo): Boolean =
-            oldItem == newItem
-
-        override fun areContentsTheSame(oldItem: AppInfo, newItem: AppInfo): Boolean =
-            oldItem.state == newItem.getCurrentState() && oldItem.selected == newItem.isNowSelected(
-                selectedList
-            )
-    }
-
-    class ViewHolder(view: View) : RecyclerView.ViewHolder(view)
-
-    interface OnItemClickListener {
-        fun onItemClick(info: AppInfo)
-    }
-
-    interface OnItemLongClickListener {
-        fun onItemLongClick(info: AppInfo): Boolean
-    }
+    override fun createFragment(position: Int): Fragment = fragments[position]
 }
